Network.AWS.SimpleDB.DomainMetadata

Description

Returns information about the domain, including when the domain was created, the number of items and attributes in the domain, and the size of the attribute names and values.

Request

data DomainMetadata

Request constructor

mkDomainMetadata

Arguments

:: Text
dmDomainName

-> DomainMetadata

Smart constructor for the minimum required parameters to construct a valid DomainMetadata request.

The fields accessible through corresponding lenses are:

DomainName :: Text

Request lenses

dmDomainName :: Lens' DomainMetadata Text

The name of the domain for which to display the metadata of.

Response

data DomainMetadataResponse

Response constructor

mkDomainMetadataResponse :: DomainMetadataResponse

Smart constructor for the minimum required parameters to construct a valid DomainMetadataResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ItemCount :: Maybe Integer

ItemNamesSizeBytes :: Maybe Integer

AttributeNameCount :: Maybe Integer

AttributeNamesSizeBytes :: Maybe Integer

AttributeValueCount :: Maybe Integer

AttributeValuesSizeBytes :: Maybe Integer

Timestamp :: Maybe Integer

Response lenses

dmrItemCount :: Lens' DomainMetadataResponse (Maybe Integer)

The number of all items in the domain.

dmrItemNamesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

The total size of all item names in the domain, in bytes.

dmrAttributeNameCount :: Lens' DomainMetadataResponse (Maybe Integer)

The number of unique attribute names in the domain.

dmrAttributeNamesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

The total size of all unique attribute names in the domain, in bytes.

dmrAttributeValueCount :: Lens' DomainMetadataResponse (Maybe Integer)

The number of all attribute name/value pairs in the domain.

dmrAttributeValuesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

The total size of all attribute values in the domain, in bytes.

dmrTimestamp :: Lens' DomainMetadataResponse (Maybe Integer)

The data and time when metadata was calculated, in Epoch (UNIX) seconds.