Network.AWS.SimpleDB.V2009_04_15.DeleteAttributes

Description

Deletes one or more attributes associated with an item. If all attributes of the item are deleted, the item is deleted. If DeleteAttributes is called without being passed any attributes or values specified, all the attributes for the item are deleted. DeleteAttributes is an idempotent operation; running it multiple times on the same item or attribute does not result in an error response. Because Amazon SimpleDB makes multiple copies of item data and uses an eventual consistency update model, performing a GetAttributes or Select operation (read) immediately after a DeleteAttributes or PutAttributes operation (write) might not return updated item data.

Request

data DeleteAttributes

Request constructor

mkDeleteAttributes

Arguments

:: Text
daDomainName

-> Text
daItemName

-> DeleteAttributes

Smart constructor for the minimum required parameters to construct a valid DeleteAttributes request.

Request lenses

daDomainName :: Lens' DeleteAttributes Text

The name of the domain in which to perform the operation.

daItemName :: Lens' DeleteAttributes Text

The name of the item. Similar to rows on a spreadsheet, items represent individual objects that contain one or more value-attribute pairs.

daAttributes :: Lens' DeleteAttributes [Attribute]

A list of Attributes. Similar to columns on a spreadsheet, attributes represent categories of data that can be assigned to items.

daExpected :: Lens' DeleteAttributes (Maybe UpdateCondition)

The update condition which, if specified, determines whether the specified attributes will be deleted or not. The update condition must be satisfied in order for this request to be processed and the attributes to be deleted.

Response

data DeleteAttributesResponse