Network.AWS.SimpleDB.V2009_04_15.DomainMetadata

Description

Returns information about the domain, including when the domain was created, the number of items and attributes in the domain, and the size of the attribute names and values.

Documentation

data DomainMetadata

Constructors

DomainMetadata

Fields

_dmrDomainName :: Text

The name of the domain for which to display the metadata of.

dmrDomainName :: Iso' DomainMetadata Text

data DomainMetadataResponse

Constructors

DomainMetadataResponse

Fields

_dmsAttributeNameCount :: Maybe Integer

The number of unique attribute names in the domain.

_dmsAttributeValueCount :: Maybe Integer

The number of all attribute name/value pairs in the domain.

_dmsItemCount :: Maybe Integer

The number of all items in the domain.

_dmsTimestamp :: Maybe Integer

The data and time when metadata was calculated, in Epoch (UNIX) seconds.

_dmsItemNamesSizeBytes :: Maybe Integer

The total size of all item names in the domain, in bytes.

_dmsAttributeValuesSizeBytes :: Maybe Integer

The total size of all attribute values in the domain, in bytes.

_dmsAttributeNamesSizeBytes :: Maybe Integer

The total size of all unique attribute names in the domain, in bytes.

dmsTimestamp :: Lens' DomainMetadataResponse (Maybe Integer)

dmsItemNamesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

dmsItemCount :: Lens' DomainMetadataResponse (Maybe Integer)

dmsAttributeValuesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

dmsAttributeValueCount :: Lens' DomainMetadataResponse (Maybe Integer)

dmsAttributeNamesSizeBytes :: Lens' DomainMetadataResponse (Maybe Integer)

dmsAttributeNameCount :: Lens' DomainMetadataResponse (Maybe Integer)