Network.AWS.SimpleDB.V2009_04_15.GetAttributes

Description

Returns all of the attributes associated with the specified item. Optionally, the attributes returned can be limited to one or more attributes by specifying an attribute name parameter. If the item does not exist on the replica that was accessed for this operation, an empty set is returned. The system does not return an error as it cannot guarantee the item does not exist on other replicas. If GetAttributes is called without being passed any attribute names, all the attributes for the item are returned.

Request

data GetAttributes

Request constructor

mkGetAttributes

Arguments

:: Text
gaDomainName

-> Text
gaItemName

-> GetAttributes

Smart constructor for the minimum required parameters to construct a valid GetAttributes request.

The fields accessible through corresponding lenses are:

DomainName :: Text

ItemName :: Text

AttributeNames :: [Text]

ConsistentRead :: Maybe Bool

Request lenses

gaDomainName :: Lens' GetAttributes Text

The name of the domain in which to perform the operation.

gaItemName :: Lens' GetAttributes Text

The name of the item.

gaAttributeNames :: Lens' GetAttributes [Text]

The names of the attributes.

gaConsistentRead :: Lens' GetAttributes (Maybe Bool)

Determines whether or not strong consistency should be enforced when data is read from SimpleDB. If true, any data previously written to SimpleDB will be returned. Otherwise, results will be consistent eventually, and the client may not see data that was written immediately before your read.

Response

data GetAttributesResponse

Response constructor

mkGetAttributesResponse :: GetAttributesResponse

Smart constructor for the minimum required parameters to construct a valid GetAttributesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Attributes :: [Attribute]

Response lenses

garAttributes :: Lens' GetAttributesResponse [Attribute]

The list of attributes returned by the operation.