Network.AWS.SimpleDB.V2009_04_15.ListDomains

Description

The ListDomains operation lists all domains associated with the Access Key ID. It returns domain names up to the limit set by MaxNumberOfDomains. A NextToken is returned if there are more than MaxNumberOfDomains domains. Calling ListDomains successive times with the NextToken provided by the operation returns up to MaxNumberOfDomains more domain names with each successive operation call.

Documentation

listDomains :: ListDomains

Minimum specification for a ListDomains request.

data ListDomains

Constructors

ListDomains

Fields

_ldrMaxNumberOfDomains :: Maybe Integer

The maximum number of domain names you want returned. The range is 1 to 100. The default setting is 100.

_ldrNextToken :: Maybe Text

A string informing Amazon SimpleDB where to start the next list of domain names.

ldrNextToken :: Lens' ListDomains (Maybe Text)

ldrMaxNumberOfDomains :: Lens' ListDomains (Maybe Integer)

data ListDomainsResponse

Constructors

ListDomainsResponse

Fields

_ldsDomainNames :: [Text]

A list of domain names that match the expression.

_ldsNextToken :: Maybe Text

An opaque token indicating that there are more domains than the specified MaxNumberOfDomains still available.

ldsNextToken :: Lens' ListDomainsResponse (Maybe Text)

ldsDomainNames :: Lens' ListDomainsResponse [Text]