Network.AWS.SimpleDB.V2009_04_15.Types

Description

Amazon SimpleDB is a highly available and flexible non-relational data store that offloads the work of database administration. Developers simply store and query data items via web services requests and Amazon SimpleDB does the rest. Unbound by the strict requirements of a relational database, Amazon SimpleDB is optimized to provide high availability and flexibility, with little or no administrative burden. Behind the scenes, Amazon SimpleDB creates and manages multiple geographically distributed replicas of your data automatically to enable high availability and data durability. The service charges you only for the resources actually consumed in storing your data and serving your requests. You can change your data model on the fly, and data is automatically indexed for you. With Amazon SimpleDB, you can focus on application development without worrying about infrastructure provisioning, high availability, software maintenance, schema and index management, or performance tuning.

Service

data SimpleDB

Supported version (2009-04-15) of the Amazon SimpleDB service.

XML

xmlOptions :: Tagged a XMLOptions

Attribute

data Attribute

mkAttribute

Arguments

:: Text
aName

-> Text
aValue

-> Attribute

Smart constructor for the minimum required fields to construct a valid Attribute data type to populate a request.

aName :: Lens' Attribute Text

The name of the attribute.

aAlternateNameEncoding :: Lens' Attribute (Maybe Text)

aValue :: Lens' Attribute Text

The value of the attribute.

aAlternateValueEncoding :: Lens' Attribute (Maybe Text)

DeletableItem

data DeletableItem

mkDeletableItem

Arguments

:: Text
diName

-> DeletableItem

Smart constructor for the minimum required fields to construct a valid DeletableItem data type to populate a request.

diName :: Lens' DeletableItem Text

diAttributes :: Lens' DeletableItem [Attribute]

Item

data Item

mkItem

Arguments

:: Text
iName

-> [Attribute]
iAttributes

-> Item

Smart constructor for the minimum required fields to construct a valid Item data type.

Item is exclusively used in responses and this constructor is provided for convenience and testing purposes.

iName :: Lens' Item Text

The name of the item.

iAlternateNameEncoding :: Lens' Item (Maybe Text)

iAttributes :: Lens' Item [Attribute]

A list of attributes.

ReplaceableAttribute

data ReplaceableAttribute

mkReplaceableAttribute

Arguments

:: Text
raName

-> Text
raValue

-> ReplaceableAttribute

Smart constructor for the minimum required fields to construct a valid ReplaceableAttribute data type to populate a request.

raName :: Lens' ReplaceableAttribute Text

The name of the replaceable attribute.

raValue :: Lens' ReplaceableAttribute Text

The value of the replaceable attribute.

raReplace :: Lens' ReplaceableAttribute (Maybe Bool)

A flag specifying whether or not to replace the attribute/value pair or to add a new attribute/value pair. The default setting is false.

ReplaceableItem

data ReplaceableItem

mkReplaceableItem

Arguments

:: Text
riName

-> [ReplaceableAttribute]
riAttributes

-> ReplaceableItem

Smart constructor for the minimum required fields to construct a valid ReplaceableItem data type to populate a request.

riName :: Lens' ReplaceableItem Text

The name of the replaceable item.

riAttributes :: Lens' ReplaceableItem [ReplaceableAttribute]

The list of attributes for a replaceable item.

UpdateCondition

data UpdateCondition

The update condition which, if specified, determines whether the specified attributes will be deleted or not. The update condition must be satisfied in order for this request to be processed and the attributes to be deleted.

mkUpdateCondition :: UpdateCondition

Smart constructor for the minimum required fields to construct a valid UpdateCondition data type to populate a request.

ucName :: Lens' UpdateCondition (Maybe Text)

The name of the attribute involved in the condition.

ucValue :: Lens' UpdateCondition (Maybe Text)

The value of an attribute. This value can only be specified when the Exists parameter is equal to true.

ucExists :: Lens' UpdateCondition (Maybe Bool)

A value specifying whether or not the specified attribute must exist with the specified value in order for the update condition to be satisfied. Specify true if the attribute must exist for the update condition to be satisfied. Specify false if the attribute should not exist in order for the update condition to be satisfied.