Network.AWS.SimpleDB.V2009_04_15.GetAttributes

Description

Returns all of the attributes associated with the specified item. Optionally, the attributes returned can be limited to one or more attributes by specifying an attribute name parameter. If the item does not exist on the replica that was accessed for this operation, an empty set is returned. The system does not return an error as it cannot guarantee the item does not exist on other replicas. If GetAttributes is called without being passed any attribute names, all the attributes for the item are returned.

Documentation

getAttributes

Arguments

:: Text
_garItemName

-> Text
_garDomainName

-> GetAttributes

Minimum specification for a GetAttributes request.

data GetAttributes

Constructors

GetAttributes

Fields

_garItemName :: Text

The name of the item.

_garDomainName :: Text

The name of the domain in which to perform the operation.

_garAttributeNames :: [Text]

The names of the attributes.

_garConsistentRead :: Maybe Bool

Determines whether or not strong consistency should be enforced when data is read from SimpleDB. If true, any data previously written to SimpleDB will be returned. Otherwise, results will be consistent eventually, and the client may not see data that was written immediately before your read.

garItemName :: Lens' GetAttributes Text

garDomainName :: Lens' GetAttributes Text

garConsistentRead :: Lens' GetAttributes (Maybe Bool)

garAttributeNames :: Lens' GetAttributes [Text]

data GetAttributesResponse

Constructors

GetAttributesResponse

Fields

_gasAttributes :: [Attribute]

The list of attributes returned by the operation.

gasAttributes :: Iso' GetAttributesResponse [Attribute]