Network.AWS.SimpleDB.V2009_04_15.ListDomains

Description

The ListDomains operation lists all domains associated with the Access Key ID. It returns domain names up to the limit set by MaxNumberOfDomains. A NextToken is returned if there are more than MaxNumberOfDomains domains. Calling ListDomains successive times with the NextToken provided by the operation returns up to MaxNumberOfDomains more domain names with each successive operation call.

Request

data ListDomains

Request constructor

mkListDomains :: ListDomains

Smart constructor for the minimum required parameters to construct a valid ListDomains request.

Request lenses

ldMaxNumberOfDomains :: Lens' ListDomains (Maybe Integer)

The maximum number of domain names you want returned. The range is 1 to 100. The default setting is 100.

ldNextToken :: Lens' ListDomains (Maybe Text)

A string informing Amazon SimpleDB where to start the next list of domain names.

Response

data ListDomainsResponse

Response constructor

mkListDomainsResponse :: ListDomainsResponse

Smart constructor for the minimum required parameters to construct a valid ListDomainsResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

ldrDomainNames :: Lens' ListDomainsResponse [Text]

A list of domain names that match the expression.

ldrNextToken :: Lens' ListDomainsResponse (Maybe Text)

An opaque token indicating that there are more domains than the specified MaxNumberOfDomains still available.